configGenerator: Bugfix,allow -saveDescriptors without real credentials (x:x) if -noUpload is active

+++ tools/config/src/com/digi/connector/config/Descriptors.java
@@ -548,7 +548,7 @@ public class Descriptors {
         if(!ConfigGenerator.noUploadOption())
             ConfigGenerator.debug_log("Uploading descriptor:" + descName);
 
-        if (callDeleteFlag) {
+        if (callDeleteFlag && !ConfigGenerator.noUploadOption()) {
             deleteDescriptors();
             callDeleteFlag = false;
         }
